Big update (make.sh; unset; redo)
- Reworking scripts, libs after adding unset, pipefail to shell options.
- Fixes for new make.sh setup, some docs
- Reworking redo build scripts

+++ doc.lib.sh
@@ -4,7 +4,6 @@
 
 doc_lib_load()
 {
-  lib_load match || return
   test -n "${DOC_EXT-}" || DOC_EXT=.rst
   test -n "${DOC_EXTS-}" || DOC_EXTS=".rst .md .txt .feature .html .htm"
   test -n "${DOC_MAIN-}" || DOC_MAIN="ReadMe main ChangeLog index doc/main docs/main"
@@ -13,6 +12,7 @@ doc_lib_load()
 doc_lib_init()
 {
   test "${doc_lib_init-}" = "0" && return
+  lib_assert match || return
 
   test -n "${package_log_doctitle_fmt-}" ||
       package_log_doctitle_fmt="%a, week %V. '%g"
@@ -21,17 +21,18 @@ doc_lib_init()
 
   test -n "${package_docs_find-}" || package_docs_find=doc_find_name
   test -n "${package_doc_find-}" || package_doc_find=doc_find_name
-
-  spwd=.
+  test -n "${package_lists_documents-}" || package_lists_documents=vc-exts-re
 
   DOC_EXTS_RE="\\$(printf -- "$DOC_EXTS" | $gsed 's/\ /\\|\\/g')"
   DOC_MAIN_RE="\\$(printf -- "$DOC_MAIN" | $gsed -e 's/\ /\\|/g' -e 's/[\/]/\\&/g')"
+  DOC_EXTS_GLOB="{**/,}*.{$( set -- $DOC_EXTS; while test $# -gt 1; do printf "${1:1},"; shift; done; printf "${1:1}"; )}"
+  DOC_EXTS_FNMATCH="$( set -- $DOC_EXTS; test $# -gt 1 && { printf -- '"*%s"' $1; shift; }; printf -- ' "*%s"' $@; )"
 }
 
 doc_path_args()
 {
   paths=$HTDIR
-  test "$(pwd)" = "$HTDIR" || {
+  test "$PWD" = "$HTDIR" || {
     paths="$paths ."
   }
 }
@@ -72,29 +73,49 @@ doc_find_name()
   find_="-false $(for ext in $DOC_EXTS ; do printf -- " -o -iname '*$ext'" ; done )"
 
   # XXX: doc-find_path_args
-  htd_find $(pwd) "$find_"
+  htd_find $PWD "$find_"
 }
 
 doc_grep_content()
 {
   test -n "$1" || set -- .
   htd_grep_excludes
-  match_grep_pattern_test "$(pwd)/"
+  match_grep_pattern_test "$PWD/"
   eval "grep -SslrIi '$1' $paths $grep_excludes" \
     | sed 's/'$p_'//'
 }
 
+# Execute package-lists-documents handler
 doc_list_local()
 {
-  test -n "$package_lists_documents" ||
-      package_lists_documents=doc-list-files-exts-re
-  upper=0 mkvid "$package_lists_documents"
-  ${vid}
+  $LOG debug "" "Doc list local..." "$package_lists_documents"
+  doc_list ${package_lists_documents} "$@"
 }
 
-doc_list_files_exts_re()
+# List all tracked files, filtering by DOC_EXTS_RE.
+# Default doc-list-local handler (package-lists-documents setting)
+doc_list ()
 {
-  spwd=. vc_tracked | grep '\('"$DOC_EXTS_RE"'\)$'
+  local srcset=${1:-"default"} ; shift
+
+  case "$srcset" in
+
+      vc-exts-re ) vc_tracked "$@" | grep '\('"$DOC_EXTS_RE"'\)$' ;;
+
+      # Does not work with Git
+      vc-exts-glob )
+           test $# -gt 0 || set -- $DOC_EXTS_GLOB
+           files_list tracked "$@"
+        ;;
+
+      # Works with: Git
+      vc-exts-fnmatch )
+           test $# -gt 0 || eval set -- $DOC_EXTS_FNMATCH
+           files_list tracked "$@"
+        ;;
+
+      * ) error "doc-list: Unknown src-set name '$srcset'" 1 ;;
+  esac
 }
 
 doc_main_files()
@@ -108,7 +129,7 @@ doc_main_files()
         test ! -e $z$x || printf -- "$z$x\\n"
       done
     done
-  done
+  done | remove_dupes
 }
 
 # Go over documents and cksums, and remove file if its md5sum matches exactly.
@@ -118,22 +139,31 @@ doc_main_files()
 htd_doc_cleanup_generated()
 {
   foreach "$cksums" | {
-      while test $# -gt 0
-      do
-        read cksum || error "No cksums left for '$1'" 1
-        { fnmatch "<*>" "$cksum" || test ! -e "$1" ; } && { shift ; continue ; }
+      local f cksum new_ck
+      while test $# -gt 0; do read cksum || {
+            error "No cksums left for '$1'" ; return 1 ; }
+
+# Ignore <*> tags if file does not exist by now
+        test -e "$1" || { shift ; continue ; }
+        $LOG debug "" "Checking against $cksum..." "$1"
+
+        fnmatch "<*>" "$cksum" && {
+            shift; continue
+        } || {
 # allow editor to work on symbolic date paths but check-in actual file
-        f="$(realpath "$1")"
-        new_ck="$(md5sum "$f" | cut -f 1 -d ' ')"
+            f="$(realpath "$1")"
+            new_ck="$(ck_md5 "$f")"
+        }
 
         test "$cksum" = "$new_ck" && {
 # Remove unchanged generated file, if not added to git
           git ls-files --error-unmatch "$f" >/dev/null 2>&1 || {
             rm "$f"
-            note "Removed unchanged generated file ($f)"
+            $LOG note "" "Removed unchanged generated file" "$f"
           }
         } || {
           git add "$f"
+          $LOG note "" "Staged" "$f"
         }
         shift
       done
@@ -196,7 +226,7 @@ args_to_title() # Title-Descr
   c=0
   { while test $# -gt 0
     do
-      { context_exists "$1" || context_existsub "$1"
+      { context_exists_tag "$1" || context_exists_subtagi "$1"
       } && {
         shift
         continue
@@ -221,8 +251,8 @@ doc_title_id() # Title-Descr...
   doc_id="$(args_to_filename "$@")"
   note "Doc-Title-Id: Doc-Id: $doc_id"
   tags="$tags $( for t in "$@" ; do
-          context_exists "$t" && echo "@$t" || {
-          context_existsub "$t" && {
+          context_exists_tag "$t" && echo "@$t" || {
+          context_exists_subtagi "$t" && {
               context_subtag_env "$t" ; echo "@$tagid" ; } ; }
           continue ; done | lines_to_words | normalize_ws_str )"
